Chance render API bump to version 6, make renderer high dpi aware.
This ports the high dpi awareness stuff from the GL3 renderer to the
Vulkan renderer. The approach is the same: Implement a new function
`SDL_Vulkan_GetDrawableSize()` and hook it into the extended renderer
API, the client uses to determine the actual drawable size which is
communicated back into the renderer. Like GL3 Vulkan has a second
representation of the drawable size in the `vid` struct, overwrite
that with the actual drawable size.

Note: Starting with this commit an up to date client which implements
renderer API version 6 is needed. Older clients won't load the renderer.